Add files via upload
It's been my pleasure to work on some interesting projects and this porject as the name suggests "Wikipedia Summary Extractor" is a Python-based tool that retrieves concise summaries of Wikipedia articles using web scraping. Unlike API-based methods, this project directly fetches and parses raw HTML content using requests and BeautifulSoup. The extracted summaries are saved in a structured JSON file for later reference.